// web/src/api/generated/models/DashboardSharedStorySummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DashboardSharedStorySummary
 */
export interface DashboardSharedStorySummary {
    /**
     * 
     * @type {number}
     * @memberof DashboardSharedStorySummary
     */
    heartMoments: number;
    /**
     * 
     * @type {number}
     * @memberof DashboardSharedStorySummary
     */
    memories: number;
    /**
     * 
     * @type {number}
     * @memberof DashboardSharedStorySummary
     */
    milestones: number;
}

/**
 * Check if a given object implements the DashboardSharedStorySummary interface.
 */
export function instanceOfDashboardSharedStorySummary(value: object): value is DashboardSharedStorySummary {
    if (!('heartMoments' in value) || value['heartMoments'] === undefined) return false;
    if (!('memories' in value) || value['memories'] === undefined) return false;
    if (!('milestones' in value) || value['milestones'] === undefined) return false;
    return true;
}

export function DashboardSharedStorySummaryFromJSON(json: any): DashboardSharedStorySummary {
    return DashboardSharedStorySummaryFromJSONTyped(json, false);
}

export function DashboardSharedStorySummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): DashboardSharedStorySummary {
    if (json == null) {
        return json;
    }
    return {
        
        'heartMoments': json['heartMoments'],
        'memories': json['memories'],
        'milestones': json['milestones'],
    };
}

export function DashboardSharedStorySummaryToJSON(json: any): DashboardSharedStorySummary {
    return DashboardSharedStorySummaryToJSONTyped(json, false);
}

export function DashboardSharedStorySummaryToJSONTyped(value?: DashboardSharedStorySummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'heartMoments': value['heartMoments'],
        'memories': value['memories'],
        'milestones': value['milestones'],
    };
}